pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.0.3",
    "icecream>=2.1.5",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.29",
    "tavily-python>=0.7.10",
]
